+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.3.0-alpha2
 
+ - Updated to mapnik-vector-tile@0.3.5 which triggers VT layers without features to no longer be encoded
  - Every layer in a vector tile that matches a layer in the mapnik.Map is now rendered and not just the first (#213)
 
 ## 1.3.0-alpha